First successful unary operator tests
