+++ wwtp_configuration/tests/data/merged.json
@@ -34,7 +34,8 @@
     "ElectricityProductionByGasVolume": {
       "tags": ["ElectricityGeneration", "GrossGasProduction"],
       "binary_operations": ["/"],
-      "type": "Efficiency"
+      "type": "Efficiency",
+      "parent_id": "Cogenerator"
     }
   },
   "DrinkingWaterFacility": {
@@ -755,6 +756,13 @@
         "units": "SCFM",
         "contents": "NaturalGas",
         "totalized": false
+      }
+    },
+    "virtual_tags": {
+      "NoGasPurchases": {
+        "tags": ["GasPurchases"],
+        "unary_operations": ["~"],
+        "type": "RunStatus"
       }
     }
   },
